split off interface trait into separate crate
