+++ Cargo.toml
@@ -19,7 +19,7 @@ is-it-maintained-open-issues = { repository = "aatxe/irc" }
 
 
 [workspace]
-members = ["./", "irc-proto/"]
+members = ["./", "irc-proto/", "irc-interface/"]
 
 
 [features]
@@ -48,13 +48,13 @@ chrono = "0.4.0"
 encoding = "0.2.0"
 futures-util = { version = "0.3.0", default-features = false, features = ["alloc", "sink"] }
 irc-proto = { version = "0.15.0", path = "irc-proto" }
+irc-interface = { version = "0.0.1", path = "irc-interface" }
 log = "0.4.0"
 parking_lot = "0.11.0"
 thiserror = "1.0.0"
 pin-project = "1.0.2"
 tokio = { version = "1.0.0", features = ["net", "time", "sync"] }
 tokio-stream = "0.1.0"
-tokio-util = { version = "0.6.0", features = ["codec"] }
 
 # Feature - Config
 serde = { version = "1.0.0", optional = true }
